Add a runserver.py script to the repository root
There is a fmn/web/main.py module to start the web application in a
development server. However, it's not configurable without editing the
file and it also is not particularly useful to anyone except for
developers. It now prints a deprecation notice and a new
``runserver.py`` script has been added to the repository root for
developer use.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>